NAFNet: fp16 inference + fp32 SCA pool (fixes 4K NaN)
Perf: run inference in fp16 (weights are fp16; fp32 input forced a full fp32
forward at ~2× cost + memory). But a real 4K optimizer run (#40 subset) then
scored VMAF 3.17 / 68 KB garbage: SCA's global average-pool sums H×W values,
overflowing fp16's ~65504 ceiling at ≥540×960 → NaN. Invisible at 128² (unit
tests) and in fp32 (parity). Fix: reduce that pool in fp32, cast back — keeps
the fp16 win everywhere else. Confirmed NaN-free at 1080p + 4K; fp32 parity
unchanged (12 tests). Added an 800×600 regression guard. Post-fix subset VMAF
recovered to 98.92 (vs off 99.49).

+++ Packages/ForgeTraining/Python/models/nafnet_mlx.py
@@ -91,7 +91,10 @@ class SCA(nn.Module):
         self.conv = Conv2d(channels, channels, kernel_size=1, bias=True)
 
     def __call__(self, x: mx.array) -> mx.array:
-        pooled = mx.mean(x, axis=(1, 2), keepdims=True)  # [N,1,1,C]
+        # Pool in fp32: the H×W sum overflows fp16 (~65504) at video resolutions
+        # → NaN. Reduce in fp32, cast back. (No-op for fp32 inference; matches
+        # NAFNet.swift.) See #40.
+        pooled = mx.mean(x.astype(mx.float32), axis=(1, 2), keepdims=True).astype(x.dtype)
         return x * self.conv(pooled)
